backend/requirements.txt
fastapi
uvicorn[standard]
spacy
pydantic
python-multipart
openai==1.98.0
python-dotenv
requests
transformers
torch
spacy-transformers 